🤖 Refactor : 코드 리팩토링
1. LottoGenerator와 LottoResultCalculator는 데이터를 이용한 비즈니스 로직을 처리하기 때문에 model/service로 변경
2. 사용자의 입력을 검사하고 비즈니스 로직을 통제하는 LottoController에서 사용자 입력 유효성 검사 수행
3. 당첨 번호와 보너스 번호를 저장하는 클래스의 이름을 LottoCompany에서 PrizeNumbers로 변경
4. 모든 유효성 검사 로직을 LottoVaildator에 위치시키고 해당하는 테스트 작성
