+++ src/main/java/lotto/util/ExceptionStatus.java
