+++ src/main/java/lotto/util/Winnings.java
