+++ src/main/java/lotto/view/SystemOutputManager.java
