--- src/test/java/controller/LottoResultCalculatorTest.java
@@ -1,20 +1,0 @@
-package controller;
-
-import lotto.controller.LottoGenerator;
-import lotto.model.Customer;
-import lotto.util.ExceptionStatus;
-import lotto.view.InputManager;
-import lotto.view.OutputManager;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.mockito.Mockito.*;
-import static org.mockito.Mockito.times;
-
-public class LottoResultCalculatorTest {
-
-
-}
