+++ src/test/java/lotto/ApplicationTest.java
