+++ src/test/java/lotto/CustomerTest.java
