+++ src/test/java/service/LottoResultCalculatorTest.java
@@ -1,0 +1,8 @@
+package service;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+public class LottoResultCalculatorTest {
+
+
+}
